Update tester.sh
not counting over 5500 attempts for the max, but showing them as skipper. If I use the tester, I know they where higher and if there are only like 1 in over 2000 as in for my project, I can just ignore it. Other people don't know that if I show it on my project tho. And if somebody else has a lot of skipped when I try, I know to test a bot with the visualizer until I find some